requirements.txt
# === Core Python Runtime ===
python-dotenv==1.2.1
requests==2.32.3
orjson==3.10.7
pyyaml==6.0.2
typing-extensions==4.12.2
packaging==24.1

# === Data Processing & ML ===
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
scipy==1.14.1
tqdm==4.66.5
regex==2024.9.11

# === LangChain / LangGraph Ecosystem ===
langchain==0.3.7
langchain-core==0.3.23
langchain-community==0.3.3
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.2
langgraph==0.2.76
langgraph-checkpoint==2.1.1
langgraph-sdk==0.1.74
langsmith==0.1.125

# === Hugging Face & Transformers ===
transformers==4.52.3
sentence-transformers==2.7.0
huggingface-hub==0.32.0       # ✅ 충돌 해결 (transformers 4.52.x 요구 조건 충족)
tokenizers==0.21.0
datasets==3.2.0               # ✅ 최신 호환 버전 (4.x는 오래됨)

# === Vector Search & Embeddings ===
faiss-cpu==1.8.0.post1
rank-bm25==0.2.2

# === Serialization & Utilities ===
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.6.1
dataclasses-json==0.6.7
joblib==1.4.2
tenacity==9.0.0

# === Web & Async IO ===
httpx==0.27.2
httpcore==1.0.5
sniffio==1.3.1
yarl==1.12.1
aiohttp==3.10.10
frozenlist==1.5.0
multidict==6.0.5

# === Document / File Handling ===
python-docx==1.1.2
pillow==10.4.0
lxml==5.3.0
pymupdf==1.24.10
psutil==5.9.8
filelock==3.15.4

# === Optional: Extra integrations ===
sqlalchemy==2.0.34
tavily-python==0.7.12

# === PyTorch (CUDA 12.4 build) ===
torch==2.5.1
triton==3.1.0

# === NVIDIA CUDA Runtime for Torch 2.5.1 ===
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-nvtx-cu12==12.4.127
nvidia-cusparse-cu12==12.4.1.24
nvidia-cusolver-cu12==11.6.1.9
nvidia-curand-cu12==10.3.5.147      # ✅ 반드시 이 버전
nvidia-cufft-cu12==11.2.1.3
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-nvjitlink-cu12==12.4.127
nvidia-nccl-cu12==2.21.5
nvidia-nvshmem-cu12==3.1.6

# === 기타 ===
xxhash==3.5.0
zstandard==0.22.0
